The following updates are done for Microsoft Exchange Server 2013 CP.
- Enhance setup instructions (bug 2578797)
- Add additional logging in the wrapper scripts (bug 2581023)
- Correct logic in the exchange_admin_audit_log.ps1 script (bug 2581023)

+++ content_support/Microsoft-Exchange/scripts/exchange_admin_audit_log.ps1
@@ -1,16 +1,29 @@
+
 function ex_admin_auditlog()
 {
   $file_name            = "last_run_date";
-  $date_from_file       = Import-Clixml $file_name -ErrorAction SilentlyContinue;
+  $date_from_file       = $null;
   $result               = $null;
-  $latest_run_date      = $date_from_file;
+  $latest_run_date      = $null;
   $previous_error_count = $Error.Count;
-
+  $logs_directory = ".\\logs";
+  
+  # Check last_run_date file exitansy
+  if (Test-Path $file_name) {
+    Write-Host ("Pick up script last execution date from last_run_date file.");
+	$date_from_file       = Get-Content $file_name -ErrorAction SilentlyContinue;
+	# Conver to date format
+	$date_from_file       = (Get-Date $date_from_file)
+  }
+  
+  # Store current date before getting logs
+  $last_run_date = Get-Date
+  
   # Get the appropriate entries.
   if ($date_from_file -eq $null) {
     $result = Search-AdminAuditLog -ErrorAction SilentlyContinue;
   } else {
-    $result = Search-AdminAuditLog -StartDate $date_from_file -ErrorAction SilentlyContinue;
+    $result = Search-AdminAuditLog -StartDate "$date_from_file" -ErrorAction SilentlyContinue;
   }
 
   # Handle Search-AdminAuditLog call failure.
@@ -19,37 +32,38 @@ function ex_admin_auditlog()
     return;
   }
 
-  # Verify the output directory exists.
-  $logs_directory = ".\\logs";
+  # Create logs output directory if not exists.
   if (!(Test-Path $logs_directory)) {
     $tempPath = New-Item -ItemType directory -Path $logs_directory;
   }
   
-   #deleting the file if it exists already 
-  if ( Test-Path $logs_directory\exchange_admin_audit_log.log ){
+  # Deleting the file if it exists already and new logs are collected
+  if ((Test-Path $logs_directory\exchange_admin_audit_log.log) -and ($result.length -gt 0)) {
 	Remove-Item $logs_directory\exchange_admin_audit_log.log
   }
 
-  foreach ($entry in $result) {
-    Write-Verbose "Checking $entry"
-    $originating_server = (("" + $entry.OriginatingServer).Split(" "))[0];
-    #$entryparameters = [string]::Format("OriginatingServer='{0}' Grep='{1}' Computer Name='{2}'", $entry.OriginatingServer, $originating_server, $env:computername);
-    #Out-File -FilePath ([string]::Format("{0}\\exchange_admin_audit_log.log", $logs_directory)) -Append -Force -InputObject $entryparameters -Encoding "UTF8";
-
-    if($originating_server -eq $env:computername) {
-        $output_entry = [string]::Format("{0} OriginatingServer='{1}', Caller='{2}', ObjectModified='{3}', CmdletName='{4}', CmdletParameters='{5}', ModifiedProperties='{6}', Succeeded='{7}', Error='{8}', IsValid='{9}'", ($entry.RunDate | Get-Date -format "yyyy MMM %d HH:mm:ss"), (("" + $entry.OriginatingServer).Split(" "))[0], $entry.Caller, $entry.ObjectModified, $entry.CmdletName, [string]::Join(",", $entry.CmdletParameters), [string]::Join(",", $entry.ModifiedProperties), $entry.Succeeded, $entry.Error, $entry.IsValid);
-        
-        if ($entry.RunDate -ge $latest_run_date) {
-            $latest_run_date = $entry.RunDate;
-        }
-
-        if ($entry.RunDate -gt $date_from_file) {
-            Out-File -FilePath ([string]::Format("{0}\\exchange_admin_audit_log.log", $logs_directory)) -Append -Force -InputObject $output_entry -Encoding "UTF8";
-        }
+  # Check if there are new logs.
+  if ($result.length -eq 0) {
+	Write-Host "There are no any new logs to get"
+  } else {  
+    # Formating result and write to the exchange_admin_audit_log.log file
+    foreach ($entry in $result) {
+      Write-Verbose "Checking $entry"
+      $originating_server = (("" + $entry.OriginatingServer).Split(" "))[0];
+      #$entryparameters = [string]::Format("OriginatingServer='{0}' Grep='{1}' Computer Name='{2}'", $entry.OriginatingServer, $originating_server, $env:computername);
+      #Out-File -FilePath ([string]::Format("{0}\\exchange_admin_audit_log.log", $logs_directory)) -Append -Force -InputObject $entryparameters -Encoding "UTF8";
+   
+      if($originating_server -eq $env:computername) {
+          $output_entry = [string]::Format("{0} OriginatingServer='{1}', Caller='{2}', ObjectModified='{3}', CmdletName='{4}', CmdletParameters='{5}', ModifiedProperties='{6}', Succeeded='{7}', Error='{8}', IsValid='{9}'", ($entry.RunDate | Get-Date -format "yyyy MMM %d HH:mm:ss"), (("" + $entry.OriginatingServer).Split(" "))[0], $entry.Caller, $entry.ObjectModified, $entry.CmdletName, [string]::Join(",", $entry.CmdletParameters), [string]::Join(",", $entry.ModifiedProperties), $entry.Succeeded, $entry.Error, $entry.IsValid);
+          
+          Out-File -FilePath ([string]::Format("{0}\\exchange_admin_audit_log.log", $logs_directory)) -Append -Force -InputObject $output_entry -Encoding "UTF8";
+      }
     }
   }
-
-  $latest_run_date | Export-Clixml $file_name;
+  # Updating latest_run_date file with lates exeuction date  
+  Out-File $file_name -InputObject $last_run_date.ToString();
+  
+  #Print result in output
   $result
 }
